requirements.txt
py-clob-client==0.34.6
httpx>=0.27.0
python-dotenv>=1.0.0
